feat(cubestore): Ability to control process in Node.js
