+++ packages/cubejs-docker/bin/cubestore-dev
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+
+require('/cubejs/rust/dist/cubestore-bin.js');
